Create Kennadiesproject
This is what I have so far for my project. The goal is to separate the tank into different bins and see how long the krill spend in them in different chla concentrations. I'm using google colab, and this is the link https://colab.research.google.com/drive/1OL_Th-zatTpLNQnSNHpNqg-hQO0YhQDY?usp=sharing